+++ web/app/themes/lark-child/config/flexible.php
@@ -1,0 +1,37 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Default Padding
+    |--------------------------------------------------------------------------
+    |
+    | This option defines the default padding for each Flexible Content Block.
+    | If there is no overriding padding selected within WordPress, this will
+    | remain the default.
+    |
+    */
+
+    'defaultPadding' => 'py-50 md:py-70',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Heading for SEO
+    |--------------------------------------------------------------------------
+    |
+    | Defining the very first Flexible Content Block to be a <h1> tag.
+    |
+    */
+
+    'h' => 1,
+
+    /*
+    |--------------------------------------------------------------------------
+    | For counting the number of accordions
+    |--------------------------------------------------------------------------
+    */
+
+    'uniqueId' => 1,
+
+];