+++ web/app/themes/lark-child/resources/views/components/blocks/container.blade.php
@@ -4,12 +4,30 @@ if( isset($background_image) ) {
 }
 @endphp
 <section
-  @if( isset($srcset) && ! empty($srcset) )data-background-image-srcset="{{ $srcset }}"@endif
-  @if( isset($style) )style="{{ $style }}"@endif
-  @hassub('id')id="{{ str_replace(' ', '-', preg_replace('/\s+/', ' ', strtolower(get_sub_field('id')))) }}"@endsub
-  class="fcb @if( isset($classes) ){{ $classes }}@endif
-  @hassub('padding_override'){{ 'fcb-' }}@sub('padding_override'){{ $padding }}@endsub
-  @hassub('background_colour'){{ 'fcb-' }}@sub('background_colour')@endsub
+  @if( isset($srcset) && ! empty($srcset) )
+    data-background-image-srcset="{{ $srcset }}"
+  @endif
+
+  @if( isset($style) )
+    style="{{ $style }}"
+  @endif
+
+  @hassub('id')
+    id="{{ str_replace(' ', '-', preg_replace('/\s+/', ' ', strtolower(get_sub_field('id')))) }}"
+  @endsub
+
+  class="fcb
+  @isset($classes)
+    {{ $classes }}
+  @endisset
+
+  @hassub('padding_override')
+    {{ 'fcb-' }}@sub('padding_override'){{ $defaultPadding }}
+  @endsub
+
+  @hassub('background_colour')
+    {{ 'fcb-' }}@sub('background_colour')
+  @endsub
 ">
   @hassub('container_type')
     <div class="@sub('container_type')">