+++ web/app/themes/lark-child/resources/views/flexible/blocks/donate.blade.php
@@ -2,7 +2,6 @@
 
   @component('components.blocks.container', [
     'classes' => 'fcb-donate',
-    'padding' => $default_padding,
   ])
 
   <div class="row">