+++ web/app/themes/lark-child/resources/views/flexible/blocks/gallery.blade.php
@@ -1,6 +1,5 @@
 @component('components.blocks.container', [
   'classes' => 'fcb-gallery pb-0',
-  'padding' => $default_padding,
 ])
 
   <div class="row">